+++ one/src/lib.rs
@@ -1,6 +1,7 @@
 //! Ceramic implements a single binary ceramic node.
 #![warn(missing_docs)]
 
+mod daemon;
 mod feature_flags;
 mod http;
 mod http_metrics;
@@ -8,32 +9,21 @@ mod metrics;
 mod migrations;
 mod network;
 
-use std::{env, path::PathBuf, time::Duration};
+use std::{env, path::PathBuf};
 
-use anyhow::{anyhow, Context, Result};
-use ceramic_api::{EventService as ApiEventService, InterestService as ApiInterestService};
-use ceramic_core::{EventId, Interest};
-use ceramic_event_svc::EventService;
-use ceramic_interest_svc::InterestService;
-use ceramic_kubo_rpc::Multiaddr;
-use ceramic_metrics::{config::Config as MetricsConfig, MetricsHandle};
-use ceramic_p2p::{load_identity, DiskStorage, Keychain, Libp2pConfig};
+use anyhow::{anyhow, Result};
+use ceramic_metrics::config::Config as MetricsConfig;
+use ceramic_sql::sqlite::SqlitePool;
 use clap::{Args, Parser, Subcommand, ValueEnum};
-use feature_flags::*;
 use futures::StreamExt;
+use libp2p::Multiaddr;
 use multibase::Base;
 use multihash::Multihash;
 use multihash_codetable::Code;
 use multihash_derive::Hasher;
-use recon::{FullInterests, Recon, ReconInterestProvider, Sha256a};
-use signal_hook::consts::signal::*;
 use signal_hook_tokio::Signals;
-use std::sync::Arc;
-use swagger::{auth::MakeAllowAllAuthenticator, EmptyContext};
 use tokio::{io::AsyncReadExt, sync::oneshot};
 use tracing::{debug, error, info, warn};
-
-use crate::network::Ipfs;
 
 #[derive(Parser, Debug)]
 #[command(author, version, about, long_about = None)]
@@ -45,194 +35,10 @@ struct Cli {
 #[derive(Subcommand, Debug)]
 enum Command {
     /// Run a daemon process
-    Daemon(Box<DaemonOpts>),
+    Daemon(Box<daemon::DaemonOpts>),
     /// Perform various migrations
     #[command(subcommand)]
     Migrations(migrations::EventsCommand),
-}
-
-#[derive(Args, Debug)]
-struct DaemonOpts {
-    #[command(flatten)]
-    db_opts: DBOpts,
-
-    /// Path to libp2p private key directory
-    #[arg(short, long, default_value=default_directory().into_os_string(), env = "CERAMIC_ONE_P2P_KEY_DIR")]
-    p2p_key_dir: PathBuf,
-
-    /// Bind address of the API endpoint.
-    #[arg(
-        short,
-        long,
-        default_value = "127.0.0.1:5101",
-        env = "CERAMIC_ONE_BIND_ADDRESS"
-    )]
-    bind_address: String,
-
-    /// Listen address of the p2p swarm.
-    #[arg(
-        long,
-        default_values_t = vec!["/ip4/0.0.0.0/tcp/4101".to_string(), "/ip4/0.0.0.0/udp/4101/quic-v1".to_string()],
-        use_value_delimiter = true,
-        value_delimiter = ',',
-        env = "CERAMIC_ONE_SWARM_ADDRESSES"
-    )]
-    swarm_addresses: Vec<String>,
-
-    /// External address of the p2p swarm.
-    /// These addressed are advertised to remote peers in order to dial the local peer.
-    #[arg(
-        long,
-        use_value_delimiter = true,
-        value_delimiter = ',',
-        env = "CERAMIC_ONE_EXTERNAL_SWARM_ADDRESSES"
-    )]
-    external_swarm_addresses: Vec<String>,
-
-    /// Extra addresses of peers that participate in the Ceramic network.
-    /// A best-effort attempt will be made to maintain a connection to these addresses.
-    #[arg(
-        long,
-        use_value_delimiter = true,
-        value_delimiter = ',',
-        env = "CERAMIC_ONE_EXTRA_CERAMIC_PEER_ADDRESSES"
-    )]
-    extra_ceramic_peer_addresses: Vec<String>,
-
-    /// Bind address of the metrics endpoint.
-    #[arg(
-        short,
-        long,
-        default_value = "127.0.0.1:9464",
-        env = "CERAMIC_ONE_METRICS_BIND_ADDRESS"
-    )]
-    metrics_bind_address: String,
-
-    /// When true metrics will be exported
-    #[arg(long, default_value_t = false, env = "CERAMIC_ONE_METRICS")]
-    metrics: bool,
-
-    /// When true traces will be exported
-    #[arg(long, default_value_t = false, env = "CERAMIC_ONE_TRACING")]
-    tracing: bool,
-
-    /// When true the tokio console will be exposed
-    #[cfg(feature = "tokio-console")]
-    #[arg(long, default_value_t = false, env = "CERAMIC_ONE_TOKIO_CONSOLE")]
-    tokio_console: bool,
-
-    /// Unique key used to find other Ceramic peers via the DHT
-    #[arg(long, default_value = "testnet-clay", env = "CERAMIC_ONE_NETWORK")]
-    network: Network,
-
-    /// Unique id when the network type is 'local'.
-    #[arg(long, env = "CERAMIC_ONE_LOCAL_NETWORK_ID")]
-    local_network_id: Option<u32>,
-
-    #[command(flatten)]
-    log_opts: LogOpts,
-
-    /// Specify maximum established outgoing connections.
-    #[arg(long, default_value_t = 2_000, env = "CERAMIC_ONE_MAX_CONNS_OUT")]
-    max_conns_out: u32,
-
-    /// Specify maximum established incoming connections.
-    #[arg(long, default_value_t = 2_000, env = "CERAMIC_ONE_MAX_CONNS_IN")]
-    max_conns_in: u32,
-
-    /// Specify maximum pending outgoing connections.
-    #[arg(long, default_value_t = 256, env = "CERAMIC_ONE_MAX_CONNS_PENDING_OUT")]
-    max_conns_pending_out: u32,
-
-    /// Specify maximum pending incoming connections.
-    #[arg(long, default_value_t = 256, env = "CERAMIC_ONE_MAX_CONNS_PENDING_IN")]
-    max_conns_pending_in: u32,
-
-    /// Specify maximum established connections per peer regardless of direction (incoming or
-    /// outgoing).
-    #[arg(long, default_value_t = 8, env = "CERAMIC_ONE_MAX_CONNS_PER_PEER")]
-    max_conns_per_peer: u32,
-
-    /// Specify idle connection timeout in milliseconds.
-    #[arg(
-        long,
-        default_value_t = 30_000,
-        env = "CERAMIC_ONE_IDLE_CONNS_TIMEOUT_MS"
-    )]
-    idle_conns_timeout_ms: u64,
-
-    /// Allowed CORS origins. Should include the transport e.g. https:// or http://.
-    #[arg(
-            long,
-            default_values_t = vec!["*".to_string()],
-            use_value_delimiter = true,
-            value_delimiter = ',',
-            env = "CERAMIC_ONE_CORS_ALLOW_ORIGINS"
-        )]
-    cors_allow_origins: Vec<String>,
-
-    /// Enable experimental feature flags
-    #[arg(
-        long,
-        use_value_delimiter = true,
-        value_delimiter = ',',
-        env = "CERAMIC_ONE_EXPERIMENTAL_FEATURE_FLAGS"
-    )]
-    experimental_feature_flags: Vec<ExperimentalFeatureFlags>,
-
-    /// Enable feature flags
-    #[arg(
-        long,
-        use_value_delimiter = true,
-        value_delimiter = ',',
-        env = "CERAMIC_ONE_FEATURE_FLAGS"
-    )]
-    feature_flags: Vec<FeatureFlags>,
-}
-
-/// The default storage directory to use if none is provided. In order:
-///     - `$HOME/.ceramic-one`
-///     -  `./.ceramic-one`
-fn default_directory() -> PathBuf {
-    home::home_dir()
-        .unwrap_or_else(|| PathBuf::from("."))
-        .join(".ceramic-one")
-}
-
-#[derive(Args, Debug)]
-struct DBOpts {
-    /// Path to storage directory.
-    #[arg(short, long, default_value=default_directory().into_os_string(), env = "CERAMIC_ONE_STORE_DIR")]
-    store_dir: PathBuf,
-}
-
-#[derive(Args, Debug)]
-struct LogOpts {
-    /// Specify the format of log events.
-    #[arg(long, default_value = "multi-line", env = "CERAMIC_ONE_LOG_FORMAT")]
-    log_format: LogFormat,
-}
-
-impl LogOpts {
-    fn format(&self) -> ceramic_metrics::config::LogFormat {
-        match self.log_format {
-            LogFormat::SingleLine => ceramic_metrics::config::LogFormat::SingleLine,
-            LogFormat::MultiLine => ceramic_metrics::config::LogFormat::MultiLine,
-            LogFormat::Json => ceramic_metrics::config::LogFormat::Json,
-        }
-    }
-}
-
-#[derive(ValueEnum, Debug, Clone, Default)]
-enum LogFormat {
-    /// Format log events on multiple lines using ANSI colors.
-    #[default]
-    MultiLine,
-    /// Format log events on a single line using ANSI colors.
-    SingleLine,
-    /// Format log events newline delimited JSON objects.
-    /// No ANSI colors are used.
-    Json,
 }
 
 #[derive(ValueEnum, Debug, Clone)]
@@ -290,24 +96,65 @@ impl Network {
     }
 }
 
+/// The default storage directory to use if none is provided. In order:
+///     - `$HOME/.ceramic-one`
+///     -  `./.ceramic-one`
+fn default_directory() -> PathBuf {
+    home::home_dir()
+        .unwrap_or_else(|| PathBuf::from("."))
+        .join(".ceramic-one")
+}
+
+// Shared options for configuring where data is stored.
+#[derive(Args, Debug)]
+struct DBOpts {
+    /// Path to storage directory.
+    #[arg(short, long, default_value=default_directory().into_os_string(), env = "CERAMIC_ONE_STORE_DIR")]
+    store_dir: PathBuf,
+}
+
+// Shared options for how logging is configured.
+#[derive(Args, Debug)]
+struct LogOpts {
+    /// Specify the format of log events.
+    #[arg(long, default_value = "multi-line", env = "CERAMIC_ONE_LOG_FORMAT")]
+    log_format: LogFormat,
+}
+
+impl LogOpts {
+    fn format(&self) -> ceramic_metrics::config::LogFormat {
+        match self.log_format {
+            LogFormat::SingleLine => ceramic_metrics::config::LogFormat::SingleLine,
+            LogFormat::MultiLine => ceramic_metrics::config::LogFormat::MultiLine,
+            LogFormat::Json => ceramic_metrics::config::LogFormat::Json,
+        }
+    }
+}
+
+#[derive(ValueEnum, Debug, Clone, Default)]
+enum LogFormat {
+    /// Format log events on multiple lines using ANSI colors.
+    #[default]
+    MultiLine,
+    /// Format log events on a single line using ANSI colors.
+    SingleLine,
+    /// Format log events newline delimited JSON objects.
+    /// No ANSI colors are used.
+    Json,
+}
+
 /// Run the ceramic one binary process
 pub async fn run() -> Result<()> {
     let args = Cli::parse();
     match args.command {
-        Command::Daemon(opts) => Daemon::run(*opts).await,
+        Command::Daemon(opts) => daemon::run(*opts).await,
         Command::Migrations(opts) => migrations::migrate(opts).await,
     }
 }
 
-type InterestInterest = FullInterests<Interest>;
-
 impl DBOpts {
     /// This function will create the database directory if it does not exist.
-    async fn get_database(
-        &self,
-        process_undelivered: bool,
-        validate_events: bool,
-    ) -> Result<Databases> {
+    async fn get_sqlite_pool(&self) -> Result<SqlitePool> {
         match tokio::fs::create_dir_all(&self.store_dir).await {
             Ok(_) => {}
             Err(err) => match err.kind() {
@@ -322,347 +169,11 @@ impl DBOpts {
             },
         }
         let sql_db_path = self.store_dir.join("db.sqlite3").display().to_string();
-        Self::build_sqlite_dbs(&sql_db_path, process_undelivered, validate_events).await
-    }
-
-    async fn build_sqlite_dbs(
-        path: &str,
-        process_undelivered: bool,
-        validate_events: bool,
-    ) -> Result<Databases> {
-        let sql_pool =
-            ceramic_sql::sqlite::SqlitePool::connect(path, ceramic_sql::sqlite::Migrations::Apply)
-                .await?;
-        let interest_svc = InterestService::new(sql_pool.clone());
-        let event_svc = EventService::try_new(sql_pool, validate_events).await?;
-        if process_undelivered {
-            event_svc.process_all_undelivered_events().await?;
-        }
-        info!(path, "connected to sqlite db");
-
-        Ok(Databases::Sqlite(SqliteBackend {
-            event_svc: event_svc.into(),
-            interest_svc: interest_svc.into(),
-        }))
-    }
-}
-
-enum Databases {
-    Sqlite(SqliteBackend),
-}
-struct SqliteBackend {
-    interest_svc: Arc<InterestService>,
-    event_svc: Arc<EventService>,
-}
-
-struct Daemon;
-
-impl Daemon {
-    async fn run(opts: DaemonOpts) -> Result<()> {
-        let info = Info::new().await?;
-
-        let mut metrics_config = MetricsConfig {
-            export: opts.metrics,
-            tracing: opts.tracing,
-            log_format: opts.log_opts.format(),
-            #[cfg(feature = "tokio-console")]
-            tokio_console: opts.tokio_console,
-            ..Default::default()
-        };
-        info.apply_to_metrics_config(&mut metrics_config);
-
-        // Currently only an info metric is recorded so we do not need to keep the handle to the
-        // Metrics struct. That will change once we add more metrics.
-        let _metrics = ceramic_metrics::MetricsHandle::register(|registry| {
-            crate::metrics::Metrics::register(info.clone(), registry)
-        });
-
-        // Logging Tracing and metrics are initialized here,
-        // debug,info etc will not work until after this line
-        let metrics_handle = ceramic_metrics::MetricsHandle::new(metrics_config.clone())
-            .await
-            .expect("failed to initialize metrics");
-        info!(
-            service__name = info.service_name,
-            version = info.version,
-            build = info.build,
-            instance_id = info.instance_id,
-            exe_hash = info.exe_hash,
-        );
-        debug!(?opts, "using daemon options");
-        let db = opts
-            .db_opts
-            .get_database(
-                true,
-                opts.experimental_feature_flags
-                    .contains(&ExperimentalFeatureFlags::EventValidation),
-            )
-            .await?;
-
-        // we should be able to consolidate the Store traits now that they all rely on &self, but for now we use
-        // static dispatch and require compile-time type information, so we pass all the types we need in, even
-        // though they are currently all implemented by a single struct and we're just cloning Arcs.
-        match db {
-            Databases::Sqlite(db) => {
-                Daemon::run_internal(
-                    opts,
-                    db.interest_svc.clone(),
-                    db.interest_svc,
-                    db.event_svc.clone(),
-                    db.event_svc.clone(),
-                    db.event_svc,
-                    metrics_handle,
-                )
-                .await
-            }
-        }
-    }
-
-    async fn run_internal<I1, I2, E1, E2, E3>(
-        opts: DaemonOpts,
-        interest_api_store: Arc<I1>,
-        interest_recon_store: Arc<I2>,
-        model_api_store: Arc<E1>,
-        model_recon_store: Arc<E2>,
-        bitswap_block_store: Arc<E3>,
-        metrics_handle: MetricsHandle,
-    ) -> Result<()>
-    where
-        I1: ApiInterestService + Send + Sync + 'static,
-        I2: recon::Store<Key = Interest, Hash = Sha256a> + Send + Sync + 'static,
-        E1: ApiEventService + Send + Sync + 'static,
-        E2: recon::Store<Key = EventId, Hash = Sha256a> + Send + Sync + 'static,
-        E3: iroh_bitswap::Store + Send + Sync + 'static,
-    {
-        let network = opts.network.to_network(&opts.local_network_id)?;
-
-        let store_dir = opts.db_opts.store_dir;
-        debug!(dir = %store_dir.display(), "using store directory");
-        debug!(dir = %opts.p2p_key_dir.display(), "using p2p key directory");
-
-        // Setup tokio-metrics
-        MetricsHandle::register(|registry| {
-            let handle = tokio::runtime::Handle::current();
-            let runtime_monitor = tokio_metrics::RuntimeMonitor::new(&handle);
-            tokio_prometheus_client::register(
-                runtime_monitor,
-                registry.sub_registry_with_prefix("tokio"),
-            );
-        });
-
-        let p2p_config = Libp2pConfig {
-            mdns: false,
-            bitswap_server: false,
-            bitswap_client: false,
-            kademlia: false,
-            autonat: false,
-            relay_server: false,
-            relay_client: false,
-            max_conns_out: opts.max_conns_out,
-            max_conns_in: opts.max_conns_in,
-            max_conns_pending_out: opts.max_conns_pending_out,
-            max_conns_pending_in: opts.max_conns_pending_in,
-            max_conns_per_peer: opts.max_conns_per_peer,
-            idle_connection_timeout: Duration::from_millis(opts.idle_conns_timeout_ms),
-            // Add extra ceramic peer addresses to the list of official ceramic peer addresses.
-            ceramic_peers: opts
-                .network
-                .bootstrap_addresses()
-                .into_iter()
-                .chain(
-                    opts.extra_ceramic_peer_addresses
-                        .iter()
-                        .map(|addr| addr.parse())
-                        .collect::<Result<Vec<Multiaddr>, multiaddr::Error>>()?,
-                )
-                .collect(),
-            external_multiaddrs: opts
-                .external_swarm_addresses
-                .iter()
-                .map(|addr| addr.parse())
-                .collect::<Result<Vec<Multiaddr>, multiaddr::Error>>()?,
-            listening_multiaddrs: opts
-                .swarm_addresses
-                .iter()
-                .map(|addr| addr.parse())
-                .collect::<Result<Vec<Multiaddr>, multiaddr::Error>>()?,
-            ..Default::default()
-        };
-        debug!(?p2p_config, "using p2p config");
-
-        // Load p2p identity
-        let mut kc = Keychain::<DiskStorage>::new(opts.p2p_key_dir.clone())
-            .await
-            .context(format!(
-                "initializing p2p key: using p2p_key_dir={}",
-                opts.p2p_key_dir.display()
-            ))?;
-        let keypair = load_identity(&mut kc).await?;
-        let peer_id = keypair.public().to_peer_id();
-
-        // Register metrics for all components
-        let recon_metrics = MetricsHandle::register(recon::Metrics::register);
-        let interest_svc_store_metrics =
-            MetricsHandle::register(ceramic_interest_svc::store::Metrics::register);
-        let event_svc_store_metrics =
-            MetricsHandle::register(ceramic_event_svc::store::Metrics::register);
-        let http_metrics = Arc::new(ceramic_metrics::MetricsHandle::register(
-            http_metrics::Metrics::register,
-        ));
-
-        // Create recon store for interests.
-        let interest_store = ceramic_interest_svc::store::StoreMetricsMiddleware::new(
-            interest_recon_store.clone(),
-            interest_svc_store_metrics.clone(),
-        );
-
-        let interest_api_store = ceramic_interest_svc::store::StoreMetricsMiddleware::new(
-            interest_api_store,
-            interest_svc_store_metrics.clone(),
-        );
-
-        // Create second recon store for models.
-        let model_store = ceramic_event_svc::store::StoreMetricsMiddleware::new(
-            model_recon_store.clone(),
-            event_svc_store_metrics.clone(),
-        );
-
-        let model_api_store = ceramic_event_svc::store::StoreMetricsMiddleware::new(
-            model_api_store,
-            event_svc_store_metrics,
-        );
-
-        // Construct a recon implementation for interests.
-        let recon_interest_svr = Recon::new(
-            interest_store.clone(),
-            InterestInterest::default(),
-            recon_metrics.clone(),
-        );
-
-        // Construct a recon implementation for models.
-        let recon_model_svr = Recon::new(
-            model_store.clone(),
-            // Use recon interests as the InterestProvider for recon_model
-            ReconInterestProvider::new(peer_id, interest_store.clone()),
-            recon_metrics,
-        );
-
-        let recons = Some((recon_interest_svr, recon_model_svr));
-        let ipfs_metrics =
-            ceramic_metrics::MetricsHandle::register(ceramic_kubo_rpc::IpfsMetrics::register);
-        let p2p_metrics = MetricsHandle::register(ceramic_p2p::Metrics::register);
-        let ipfs = Ipfs::<E3>::builder()
-            .with_p2p(
-                p2p_config,
-                keypair,
-                recons,
-                bitswap_block_store.clone(),
-                p2p_metrics,
-            )
-            .await?
-            .build(bitswap_block_store, ipfs_metrics)
-            .await
-            .map_err(|e| {
-                anyhow!(
-                    "Failed to start libp2p server using addresses: {}. {}",
-                    opts.swarm_addresses.join(", "),
-                    e
-                )
-            })?;
-
-        // Start metrics server
-        debug!(
-            bind_address = opts.metrics_bind_address,
-            "starting prometheus metrics server"
-        );
-        let (tx_metrics_server_shutdown, metrics_server_handle) =
-            metrics::start(&opts.metrics_bind_address.parse()?).map_err(|e| {
-                anyhow!(
-                    "Failed to start metrics server using address: {}. {}",
-                    opts.metrics_bind_address,
-                    e
-                )
-            })?;
-
-        // Build HTTP server
-        let mut ceramic_server = ceramic_api::Server::new(
-            peer_id,
-            network,
-            interest_api_store,
-            Arc::new(model_api_store),
-        );
-        if opts
-            .experimental_feature_flags
-            .contains(&ExperimentalFeatureFlags::Authentication)
-        {
-            ceramic_server.with_authentication(true);
-        }
-        let ceramic_service = ceramic_api_server::server::MakeService::new(ceramic_server);
-        let ceramic_service = MakeAllowAllAuthenticator::new(ceramic_service, "");
-        let ceramic_service =
-            ceramic_api_server::context::MakeAddContext::<_, EmptyContext>::new(ceramic_service);
-
-        let kubo_rpc_server = ceramic_kubo_rpc::http::Server::new(ipfs.api());
-        let kubo_rpc_service = ceramic_kubo_rpc_server::server::MakeService::new(kubo_rpc_server);
-        let kubo_rpc_service = MakeAllowAllAuthenticator::new(kubo_rpc_service, "");
-        let kubo_rpc_service =
-            ceramic_kubo_rpc_server::context::MakeAddContext::<_, EmptyContext>::new(
-                kubo_rpc_service,
-            );
-        let kubo_rpc_service =
-            http::MakeMetricsService::new(kubo_rpc_service, http_metrics.clone());
-        let kubo_rpc_service =
-            http::MakeCorsService::new(kubo_rpc_service, opts.cors_allow_origins.clone());
-        let ceramic_service = http::MakeCorsService::new(ceramic_service, opts.cors_allow_origins);
-        let ceramic_service = http::MakeMetricsService::new(ceramic_service, http_metrics);
-
-        // Compose both services
-        let service = http::MakePrefixService::new(
-            ("/ceramic/".to_string(), ceramic_service),
-            ("/api/v0/".to_string(), kubo_rpc_service),
-        );
-
-        // Start HTTP server with a graceful shutdown
-        let (tx, rx) = tokio::sync::oneshot::channel::<()>();
-        let signals = Signals::new([SIGHUP, SIGTERM, SIGINT, SIGQUIT])?;
-        let handle = signals.handle();
-
-        debug!("starting signal handler task");
-        let signals_handle = tokio::spawn(handle_signals(signals, tx));
-
-        // The server task blocks until we are ready to start shutdown
-        info!("starting api server at address {}", opts.bind_address);
-        hyper::server::Server::try_bind(&opts.bind_address.parse()?)
-            .map_err(|e| anyhow!("Failed to bind address: {}. {}", opts.bind_address, e))?
-            .serve(service)
-            .with_graceful_shutdown(async {
-                rx.await.ok();
-            })
-            .await?;
-        debug!("api server finished, starting shutdown...");
-
-        // Stop IPFS.
-        if let Err(err) = ipfs.stop().await {
-            warn!(%err,"ipfs task error");
-        }
-        debug!("ipfs stopped");
-
-        // Shutdown metrics server and collection handler
-        tx_metrics_server_shutdown
-            .send(())
-            .expect("should be able to send metrics shutdown message");
-        if let Err(err) = metrics_server_handle.await? {
-            warn!(%err, "metrics server task error")
-        }
-        metrics_handle.shutdown();
-        debug!("metrics server stopped");
-
-        // Wait for signal handler to finish
-        handle.close();
-        signals_handle.await?;
-        debug!("signal handler stopped");
-
-        Ok(())
+        Ok(ceramic_sql::sqlite::SqlitePool::connect(
+            &sql_db_path,
+            ceramic_sql::sqlite::Migrations::Apply,
+        )
+        .await?)
     }
 }
